Parametro funcionando

+++ fichasordens-domain/src/main/java/br/com/fichasordens/repository/ParametroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fichasordens.entities.ParametroEntity;
 
-public interface ParametroRepository extends JpaRepository<ParametroEntity, Integer> {
+public interface ParametroRepository extends JpaRepository<ParametroEntity, Byte> {
 
 }
